Remove references to Octopus Energy
This is now a personal project for ease of publishing.
