--- core/src/commonMain/kotlin/energy/octopus/monarch/FeatureFlagManager.kt
@@ -1,12 +1,0 @@
-package energy.octopus.monarch
-
-/**
- * Manage the options a [FeatureFlag] can manifest as.
- */
-public interface FeatureFlagManager {
-
-    /**
-     * Get the current value of [flag].
-     */
-    public fun <T : Any> currentValueFor(flag: FeatureFlag<T>): T
-}
