+++ core/src/commonMain/kotlin/io/github/kevincianfarini/monarch/ObservableFeatureFlagManager.kt
@@ -1,0 +1,11 @@
+package io.github.kevincianfarini.monarch
+
+import kotlinx.coroutines.flow.Flow
+
+public interface ObservableFeatureFlagManager : FeatureFlagManager {
+
+    /**
+     * Emits value changes of [flag].
+     */
+    public fun <T : Any> valuesFor(flag: FeatureFlag<T>): Flow<T>
+}
