+++ gradle.properties
@@ -1,1 +1,25 @@
 kotlin.code.style=official
+
+GROUP=io.github.kevincianfarini.monarch
+POM_ARTIFACT_ID=monarch
+VERSION_NAME=0.1.0-SNAPSHOT
+
+POM_NAME=Monarch
+POM_DESCRIPTION=Kotlin Multiplatform feature flagging
+POM_INCEPTION_YEAR=2024
+
+POM_URL=https://github.com/kevincianfarini/monarch
+POM_SCM_URL=https://github.com/kevincianfarini/monarch
+POM_SCM_CONNECTION=scm:git:git://github.com/kevincianfarini/monarch.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/kevincianfarini/monarch.git
+
+POM_LICENCE_NAME=The Apache Software License, Version 2.0
+POM_LICENCE_URL=https://www.apache.org/licenses/LICENSE-2.0.txt
+POM_LICENCE_DIST=repo
+
+POM_DEVELOPER_ID=kevincianfarini
+POM_DEVELOPER_NAME=Kevin Cianfarini
+POM_DEVELOPER_URL=https://github.com/kevincianfarini
+
+SONATYPE_CONNECT_TIMEOUT_SECONDS=120
+SONATYPE_CLOSE_TIMEOUT_SECONDS=900
