--- mixins/kotlinx-serialization-json/src/commonTest/kotlin/energy/octopus/monarch/mixins/JsonFeatureFlagTest.kt
@@ -1,27 +1,0 @@
-package energy.octopus.monarch.mixins
-
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class JsonFeatureFlagTest {
-
-    @Test fun `deserializes with supplied serializer`() {
-        val jsonString = """{"bar":2}"""
-        assertEquals(
-            expected = Foo(2),
-            actual = SomeJsonFlag.deserialize(jsonString, Json.Default)
-        )
-    }
-}
-
-object SomeJsonFlag : JsonFeatureFlag<Foo>(
-    key = "some_flag",
-    default = Foo(1),
-    deserializer = Foo.serializer(),
-)
-
-@Serializable data class Foo(
-    val bar: Int,
-)
